+++ packages/site/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- add styles for `TxModal`
 - Fix unwrap / spend amounts in `index.tsx`
 - Add tx modal to `playground.tsx`
 - Fix solidity array input bug (misinterpreted as string)